+++ barapost/src/barapost_local_modules/parallel_mult_files.py
@@ -3,17 +3,16 @@
 #   processing in "many-files" mode.
 
 import os
-import shelve
 import multiprocessing as mp
 from re import search as re_search
 
 from src.fasta import fasta_packets
 from src.fastq import fastq_packets
 
-from src.printlog import getwt, printl, printn, println
+from src.printlog import getwt, printl, println
 from src.write_classification import write_classification
 from src.spread_files_equally import spread_files_equally
-from src.filesystem import get_curr_res_dpath, create_result_directory
+from src.filesystem import create_result_directory
 from src.filesystem import remove_tmp_files, is_fastq, OPEN_FUNCS, FORMATTING_FUNCS, is_gzipped
 
 from src.barapost_local_modules.barapost_spec import look_around, launch_blastn, parse_align_results_xml
@@ -72,7 +71,7 @@ def process_paral(fq_fa_list, packet_size, tax_annot_res_dir,
     for fq_fa_path in fq_fa_list:
 
         # Create the result directory with the name of FASTQ of FASTA file being processed:
-        new_dpath = create_result_directory(fq_fa_path, tax_annot_res_dir)
+        new_dpath = create_result_directory(fq_fa_path, tax_annot_res_dir, logfile_path)
 
         # "hname" means human readable name (i.e. without file path and extention)
         infile_hname = os.path.basename(fq_fa_path)
@@ -139,7 +138,8 @@ def process_paral(fq_fa_list, packet_size, tax_annot_res_dir,
         # end with
     # end for
 
-    remove_tmp_files( os.path.join(queries_tmp_dir, "query{}_tmp.fasta".format(os.getpid())) )
+    query_fpath = os.path.join(queries_tmp_dir, "query{}_tmp.fasta".format(os.getpid()))
+    remove_tmp_files([query_fpath], logfile_path)
 # end def process_paral